[DEAD END] Add additional experimental graphql functions
This includes both `ghub--graphql' and other low-level functions that
asynchronously fetch and automatically unpagination from Github, as
well as some functions built on top of those, intended to be used by
the new `forge' package.

All of these functions are still experimental and still likely to
undergo major changes.  They are not part of the official API yet.